requirements.txt
asgiref==3.2.10
astroid==2.3.3
Babel==2.8.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
Django==3.0.7
django-cors-headers==3.2.1
django-crispy-forms==1.9.2
django-rest-swagger==2.2.0
djangorestframework==3.10.0
djangorestframework-simplejwt==4.4.0
factory-boy==3.0.1
Faker==4.1.1
gunicorn==20.0.4
idna==2.9
isort==4.3.21
itypes==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
Pillow==7.0.0
postgres==3.0.0
psycopg2-binary==2.8.4
psycopg2-pool==1.1
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.8
whitenoise==5.2.0
wrapt==1.11.2
